requirements.txt
click==7.1.1
flask-marshmallow==0.11.0
flask-sqlalchemy==2.4.1
flask==1.1.1
itsdangerous==1.1.0
jinja2==2.11.1
markupsafe==1.1.1
marshmallow-sqlalchemy==0.22.3
marshmallow==3.5.1
six==1.14.0
sqlalchemy==1.3.15
werkzeug==1.0.0
google-cloud-datastore==1.7.3
